refactor: unify signal delivery through self-pipe, fix DNS UAF
Signal subsystem rewrite:
- Remove sig_pending flag and Windows bridge code
- Unified pipe-based delivery on both platforms
  POSIX: async-signal-safe write() in signal handler
  Windows: SetConsoleCtrlHandler callback writes to TCP pipe
- ccev__signal_dispatch reads pipe and queues ccev_signal_event_t
- ccev__signal_process_queue drains queue at loop step 7

S-1 use-after-free fix (connect timeout + DNS race):
- Inline _connect_dns_ctx_t into connector as dns_alive flag
- Remove separate heap allocation (no more three-owner free race)
- _connect_timeout_cb / _connector_close_cb only set flag
- _connect_dns_cb is sole consumer

Memory allocator isolation:
- Extract ccev_set_allocator + function pointers into ccev_mem.c

Codebase audit fixes (security/reliability):
- ccev_signal_handle: store callback only after install succeeds
- Fix Windows dwCtrlType→signum mapping (was writing CTRL_C_EVENT=0)
- Fix Windows BOOL handler missing return values
- Fix ccev__console_ctrl_handler dangling reference

+++ src/ccev_dns.c
@@ -342,7 +342,7 @@ static int ccev__dns_send_one(ccev_loop_t *loop, ccsocket_t fd,
                                ccdns_t *ctx, const char *domain,
                                ccdns_type_t qtype) {
     unsigned char buf[512];
-    uint16_t qlen = ccdns_encode(ctx, buf, sizeof(buf), domain, qtype);
+    uint16_t qlen = ccdns_encode(ctx, buf, (uint16_t)sizeof(buf), domain, qtype);
     if (qlen == 0) return CCEV_ERR;
     int sent = 0;
     ccsocket_sendto(fd, (const char*)buf, qlen,
@@ -371,9 +371,9 @@ static void dns_recv_cb(ccev_sock_t *sock, int events) {
     dns_collect_t col;
     memset(&col, 0, sizeof(col));
 
-    int ret = ccdns_decode(&q->ctx_a, buf, (size_t)n, &col, dns_collect_cb);
+    int ret = ccdns_decode(&q->ctx_a, buf, (uint16_t)n, &col, dns_collect_cb);
     if (ret < 0)
-        ccdns_decode(&q->ctx_aaaa, buf, (size_t)n, &col, dns_collect_cb);
+        ccdns_decode(&q->ctx_aaaa, buf, (uint16_t)n, &col, dns_collect_cb);
 
     char addr[256] = "";
     int status = CCEV_ERR;
